Python/requirements.txt
absl-py==0.10.0
bleak==0.13.0
bluepy==1.3.0
cffi==1.14.2
cryptography==3.1
cycler==0.10.0
enum-compat==0.0.3
future==0.18.2
kiwisolver==1.1.0
loguru==0.5.2
matplotlib
numpy==1.18.5
pc-ble-driver-py==0.14.2
pexpect==4.8.0
ptyprocess==0.6.0
pycparser==2.20
pygatt==4.0.5
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
wrapt==1.12.1
